requirements.txt
fastapi == 0.116.1
pydantic == 2.11.7
tronpy == 0.6.1
SQLAlchemy == 2.0.41
aiokafka == 0.12.0
uvicorn == 0.35.0
python-dotenv == 1.1.1
psycopg2 == 2.9.10
pytest == 8.4.1
